# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Watchpost",
    "description": "Agent-native monitoring service. Create monitors, track uptime, receive alerts — all via REST API. No signup required.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/monitors/{id}/uptime": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": {
        "summary": "Get uptime statistics",
        "operationId": "getUptime",
        "tags": ["heartbeats"],
        "responses": {
          "200": { "description": "Uptime stats", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UptimeStats" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/incidents/{id}/acknowledge": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "summary": "Acknowledge an incident",
        "operationId": "acknowledgeIncident",
        "tags": ["incidents"],
        "security": [{ "manageKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AcknowledgeIncident" } } }
        },
        "responses": {
          "200": { "description": "Incident acknowledged" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/monitors/{id}/maintenance": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "post": {
        "summary": "Create a maintenance window (suppresses incidents during the window)",
        "operationId": "createMaintenanceWindow",
        "tags": ["maintenance"],
        "security": [{ "manageKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "starts_at", "ends_at"],
                "properties": {
                  "title": { "type": "string", "description": "Maintenance description" },
                  "starts_at": { "type": "string", "format": "date-time", "description": "Window start (ISO-8601 UTC)" },
                  "ends_at": { "type": "string", "format": "date-time", "description": "Window end (ISO-8601 UTC)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Maintenance window created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MaintenanceWindow" } } } },
          "400": { "description": "Validation error" },
          "403": { "description": "Invalid manage key" }
        }
      },
      "get": {
        "summary": "List maintenance windows for a monitor",
        "operationId": "listMaintenanceWindows",
        "tags": ["maintenance"],
        "responses": {
          "200": { "description": "List of maintenance windows", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MaintenanceWindow" } } } } },
          "404": { "description": "Monitor not found" }
        }
      }
    },
    "/maintenance/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "delete": {
        "summary": "Delete a maintenance window",
        "operationId": "deleteMaintenanceWindow",
        "tags": ["maintenance"],
        "security": [{ "manageKey": [] }],
        "responses": {
          "200": { "description": "Deleted" },
          "403": { "description": "Invalid manage key" },
          "404": { "description": "Maintenance window not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "manageKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Monitor manage_key. Also accepted as X-API-Key header or ?key= query param."
      },
      "adminKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Service admin key (auto-generated on first run, printed to stdout). Used for service-level settings."
      }
    },
    "schemas": {
...
      "AcknowledgeIncident": {
        "type": "object",
        "required": ["note"],
        "properties": {
          "note": { "type": "string" },
          "actor": { "type": "string", "default": "anonymous" }
        }
      },
      "UptimeStats": {
        "type": "object",
        "properties": {
          "monitor_id": { "type": "string" },
          "uptime_24h": { "type": "number" },
          "uptime_7d": { "type": "number" },
          "uptime_30d": { "type": "number" },
          "uptime_90d": { "type": "number" },
          "total_checks_24h": { "type": "integer" },
          "total_checks_7d": { "type": "integer" },
          "total_checks_30d": { "type": "integer" },
          "total_checks_90d": { "type": "integer" },
          "avg_response_ms_24h": { "type": "number", "nullable": true }
        }
      },
...
      "MaintenanceWindow": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "monitor_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "starts_at": { "type": "string", "format": "date-time" },
          "ends_at": { "type": "string", "format": "date-time" },
          "active": { "type": "boolean", "description": "Whether the window is currently active" },
          "created_at": { "type": "string" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "Invalid manage key",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}